Completed DestiniRPG App
This app is a storyline sequence based game that allows you to choose
your destiny (based on decisions made). This app focused on logical
statements to navigate the player from story to story. I added a
restart button at the end of each dead-end sequence that allows players
to start from the beginning to “change their path”.